% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcinfo_mixsiar.R
\name{calcinfo_mixsiar}
\alias{calcinfo_mixsiar}
\title{Functions to calculate information divergence measures for mixsiar models}
\usage{
calcinfo_mixsiar(mixmodlist, prior.control = NULL, metric = c("hellinger",
  "kldivergence"))
}
\arguments{
\item{mixmodlist}{a \code{list} with a simmr_input and simmr_output object.}

\item{dp}{the prior used, generated with function default_prior
, if NULL, assumes the default.}
}
\value{
Relevant information or evaluation measure,
assuming default priors for simmr.
}
\description{
Functions to calculate information divergence measures for mixsiar models
}
\details{
take an object from \code{simmr_fitsims} and calculates and
returns evaluation metrics.
}
\author{
Christopher J. Brown
}
