% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calckl.R
\name{calckl}
\alias{calckl}
\title{Functions to calculate evaluation measures}
\usage{
calckl(mod_run, dp = NULL)
}
\arguments{
\item{mod_run}{a \code{list} with a simmr_input and simmr_output object.}

\item{dp}{the prior used, generated with function default_prior
, if NULL, assumes the default.}
}
\value{
Relevant information or evaluation measure,
assuming default priors for simmr.
}
\description{
Functions to calculate evaluation measures
}
\details{
take an object from \code{simmr_fitsims} and calculates and
returns evaluation metrics.
}
\author{
Christopher J. Brown
}
