% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_prior.R
\name{default_prior}
\alias{default_prior}
\title{Default priors for the simmr package}
\usage{
default_prior(simmr_in)
}
\arguments{
\item{simmr_in}{A \code{simmr_input} object.}
}
\value{
Prior means and standard deviations.
\item{means}{Prior means}
\item{sd}{Standard deviations}
}
\description{
Default priors for the simmr package
}
\author{
Christopher J. Brown
}
