% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpriors.R
\name{drawpriors}
\alias{drawpriors}
\title{Random draws from simmr prior distributions}
\usage{
drawpriors(simmr_in, priorcontrol, ndraws = 50000, plot.priors = FALSE,
  cor.table = FALSE)
}
\arguments{
\item{simmr_in}{A \code{simmr_input} object with the data for input
to a simmr model}

\item{priorcontrol}{A list with the means and standard deviations of the
priors.}

\item{ndraws}{A \code{numeric} giving the number of draws from the prior
distributions.}

\item{plot.priors}{A \code{logical} should the prior densities be plotted?}

\item{cor.table}{A \code{logical} should a table of correlations among
sources be printed?}
}
\value{
A \code{data.frame} of random draws from the priors.
}
\description{
Random draws from simmr prior distributions
}
\details{
Newer isotope mixing models, including simmr, use logistic-normal
transformations. So \code{priorcontrol} should give means and sds in normal
space. If you want to specify priors as proportions see:
\code{simmr::simmr_elicit}
}
\author{
Christopher J. Brown christo.j.brown@gmail.com
}
