% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gulls}
\alias{gulls}
\title{Gulls data-set}
\format{list:
\describe{
  \item{mixes}{Carbon and nitrogen values for gulls}
  \item{s_names}{source names}
  \item{source_means}{Mean carbon and nitrogen values
 for the sources}
  \item{sources_sds}{Standard deviation of the 
 carbon and nitrogen values for the sources}
}}
\usage{
gulls
}
\description{
This dataset is simulations of data based on the findings in
Masello JF, Wikelski M, Voigt CC, Quillfeldt P. 
Distribution patterns predict individual specialization
in the diet of Dolphin Gulls. PLoS One. 2013 Jul 2:e67714.
The sample size for gulls has been increased to 50.
}
\keyword{datasets}
