% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gulls_out}
\alias{gulls_out}
\title{Output of simmr analysis on cloned gull data}
\format{simmr_output}
\usage{
gulls_out
}
\description{
Output of simmr analysis on cloned gull data
}
\keyword{datasets}
