% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gulls_poly}
\alias{gulls_poly}
\title{Gulls outside the polygon data-set}
\format{list:
\describe{
  \item{mixes}{Carbon and nitrogen values for gulls}
  \item{s_names}{source names}
  \item{source_means}{Mean carbon and nitrogen values
 for the sources}
  \item{sources_sds}{Standard deviation of the 
 carbon and nitrogen values for the sources}
}}
\usage{
gulls_poly
}
\description{
This dataset is based on the gulls_small dataset
but nitrogen measurements for gulls have been shifted
so they fall outside the mixing polygon.
}
\keyword{datasets}
