% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gulls_small}
\alias{gulls_small}
\title{Gulls data-set with small sample size}
\format{list:
\describe{
  \item{mixes}{Carbon and nitrogen values for gulls}
  \item{s_names}{source names}
  \item{source_means}{Mean carbon and nitrogen values
 for the sources}
  \item{sources_sds}{Standard deviation of the 
 carbon and nitrogen values for the sources}
}}
\usage{
gulls_small
}
\description{
This dataset is based simulations of data based on the findings in
Masello JF, Wikelski M, Voigt CC, Quillfeldt P. 
Distribution patterns predict individual specialization
in the diet of Dolphin Gulls. PLoS One. 2013 Jul 2:e67714.
}
\keyword{datasets}
