% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixsiar_fitsims.R
\name{mixsiar_fitsims}
\alias{mixsiar_fitsims}
\title{Fits a MixSIAR model to a randomly generated data-set}
\usage{
mixsiar_fitsims(n, tracer_means, tracer_sds, snames, modelfile, mixsiardat,
  sdmult = 1, seed = 42, mcmc.control = list(iter = 50000, burn = 10000,
  thin = 10, n.chain = 4), prior.control = NULL)
}
\arguments{
\item{n}{A \code{integer} giving the sample size to simulate.}

\item{tracer_means}{A \code{numeric} vector giving means of the mixes.}

\item{tracer_sds}{A \code{numeric} vector giving SDs of the mixes.}

\item{snames}{A \code{character} giving source names.}

\item{modelfile}{A \code{character} giving the location of the MixSIAR
model filename.}

\item{mixsiardat}{A \code{list} giving MixSIAR input data that will be
modified for the re-run (should include objects mix, source, and disc, see
the MixSIAR manual for further info).}

\item{sdmult}{A \code{numeric} that is multiplied by source SDs}

\item{mcmc.control}{A list giving parameters for the MCMC chains.}

\item{priorcontrol}{A list with the means and standard deviations of the
priors.}
}
\value{
A list where the first object gives the simulated data and the
second object is a \code{rjags} object giving the model fit.
}
\description{
Useful for obtaining fits from a sequence of randomly simulated datasets.
For instance, you could use this function to perform a sort of broad-sense
power analysis to ask how many samples are required to obtain a specific
confidence about an animal's diet.
}
\details{
The mixes and sampled randomly from normal distributions.
Fixing the seed is useful if you want to compare runs across different
SDs, because the random samples will have the same pattern deviations.
}
\author{
Christopher J. Brown
}
