% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pianka.R
\name{pianka}
\alias{pianka}
\title{Calculate Pianka's similarity measure on sources}
\usage{
pianka(nsources, s_means, s_sds)
}
\arguments{
\item{nsources}{\code{integer} giving number of sources}

\item{s_means}{\code{matrix} giving tracer means (columns) for each source}

\item{s_sds}{\code{matrix} giving tracer SD (columns) for each source}
}
\value{
\code{matrix} giving Pianka's similarity measure.
}
\description{
Calculate Pianka's similarity measure on sources
}
\details{
Calculates Pianka's measure assuming isotope values of sources
are multivariate normal. Note that equation as given in Lu et al. 1989
returns a square matrix, so I am taking value (1,1).
}
\author{
Christopher J. Brown
}
