% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dists.R
\name{plot_dists}
\alias{plot_dists}
\title{Calculates Hellinger distance and plot prior and posterior distributions}
\usage{
plot_dists(simmr_in, simmr_out, priorcontrol, plotdist = FALSE)
}
\arguments{
\item{simmr_in}{A \code{simmr_input} object.}

\item{simmr_out}{A \code{simmr_output} object.}

\item{priorcontrol}{A list with the means and standard deviations of the
priors.}

\item{plotdist}{A logical or integer vector indicating whether to plot the distributions. If an integer, then the group(s) to plot.}
}
\value{
A \code{remix_dists} object that contains the Hellinger distances.
\item{hellinger_discrete}{Hellinger distance estimated using a discrete
approximation of the densities}
\item{hellinger_continuous}{Hellinger distance estimated using a continuous
approximation of the densities}
\item{distfits}{A list for \code{BayeSens::helldist} objects containing details of the
functions fitted to the densities.}
}
\description{
This function calculates the Hellinger distance between prior and posterior
It can also plot the distributions.
}
\details{
Values are numerical approximations. See \code{BayeSens::hellinger} for more
details on estimation of Hellinger distances from MCMC chains.
In general the continuous estimation will be more accurate, however large
discrepencies between discrete and continuous estimates suggest bi-modality,
in which case we recommned the discrete estimate be used.
Items of the \code{distfits} list can be plotted to check distribution fits.
}
\author{
Christopher J. Brown
}
