% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_chains.R
\name{post_chains}
\alias{post_chains}
\title{Extract chains for the posterior distribution}
\usage{
post_chains(simmr_out, plot.post = FALSE)
}
\arguments{
\item{simmr_out}{A \code{simmr_output} object.}

\item{plot.post}{A logical or integer vector indicating whether to plot the
distributions. If an integer, then the group(s) to plot.}
}
\value{
A list that contains data.frames of the MCMC draws for each group.
}
\description{
Extract chains for the posterior distribution
}
\details{
Normally, this function need not be called directly.
Instead use \code{\link{plot_dists}} to calculate Hellinger distances
and plot posterior distributions.
}
\author{
Christopher J. Brown
}
