% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix_2DPI.R
\name{remix_2DPI}
\alias{remix_2DPI}
\title{Isospace plots of model predicted consumer isotope ratios}
\usage{
remix_2DPI(x2, simmr_in, groups = NULL, tracerpairs = NULL, nxvals = 200,
  xlab = colnames(simmr_in$mixtures)[1],
  ylab = colnames(simmr_in$mixtures)[2], prob = c(0.05, 0.25),
  xlims = NULL, ylims = NULL, plotCI = TRUE)
}
\arguments{
\item{x2}{A \code{remix_simdata} object.}

\item{simmr_in}{A \code{simmr_input} object.}

\item{groups}{A \code{numeric} with the groups to plot if they are used,
otherwise \code{NULL}.}

\item{tracerpairs}{A two row \code{matrix} were columns are the tracer pairs
you want to plot.}

\item{nxvals}{A \code{numeric} giving the number of intervals on the
x and y axes.}

\item{xlab}{A \code{character} the x-axis label}

\item{ylab}{A \code{character} the y-axis label}

\item{prob}{A \code{numeric} vector giving the probability intervals
to plot.}

\item{xlims}{A \code{numeric} pair giving the x-axis limits}

\item{ylims}{A \code{numeric} pair giving the y-axis limits}

\item{plotCI}{A \code{logical} which determines whether credibility interavls
are plotted in addition to the predictive intervals.}
}
\value{
Plots of the tracer pairs with predicted isotope
values of consumers.
}
\description{
Plotting the consumer's isotope values against those predicted by the model
can be useful to detect model bias, or instance when consumers fall outside
the source polygon.
}
\details{
Plots the output of \code{\link{remix_simdata}} on two axes. remix_2DPI also
plots predictive intervals at the given level, which are spaces where the
consumer isotope ratios have a 95% (or other probability interval) of
falling. Optionally can also plot 95% (or other interval) credibility
intervals, which indicate a 95% probability of the consumer's mean value.
Model bias is indicated by consumer data that fall outside the predictive
intervals. Bias may occur, among other things due to
missed sources or incorrect fractionation estimates.
}
\author{
Christopher J. Brown
}
