% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix_pca.R
\name{remix_pca}
\alias{remix_pca}
\title{Perform a principal components analysis on posterior chains}
\usage{
remix_pca(simmr_out, transform = F)
}
\arguments{
\item{simmr_out}{A \code{simmr_output} object.}

\item{transform}{A \code{logical} indicating whether to transform diet proportions
using the isometric log-ratio transform.}
}
\value{
A \code{prcomp} object
}
\description{
PCA can be used as a tool to visualise correlations among
estimates of source contributions.
}
\details{
See \code{prcomp} for more details. \code{remix_pca} has a print method, to view the variance components. Use \code{remix_biplot} to view a biplot. 
Transforming the data will tend to increase emphasis on correlations among smaller values.
}
\author{
Christopher J. Brown
}
