% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix_pcaplot.R
\name{remix_pcaplot}
\alias{remix_pcaplot}
\title{Plot the axes loadings from a PCA}
\usage{
remix_pcaplot(pc, axes = c(1, 2))
}
\arguments{
\item{pc}{A \code{prcomp} object.}

\item{axes}{A \code{numeric} vector giving the two 
axes to plot.}
}
\description{
Plots vectors representing the loadings of each source
on two PCA axes.
}
\details{
This function is similar to a PCA biplot, except that the 
observations are not plotted. This speeds up plotting for
when a PCA has been run on MCMC chains, which may have
100 000s of observations.
}
\author{
Christopher J. Brown
}
