% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix_PI.R
\name{remix_PI}
\alias{remix_PI}
\title{Estimate predictive intervals for consumer isotope ratios}
\usage{
remix_PI(simdat, simmr_in, simmr_out, groups = NULL, plotresid = TRUE,
  probs = c(0.025, 0.975))
}
\arguments{
\item{simmr_in}{A \code{simmr_input} object.}

\item{groups}{A \code{numeric} with the groups to plot if they are used,
otherwise \code{NULL}.}

\item{plotresid}{A \code{logical} which determines whether credibility
residual plots are created.}

\item{probs}{A \code{numeric} vector of two numbers giving the upper and
lower predictive intervals.}

\item{simdata}{A \code{remix_PI} object.}

\item{simmr_in}{A \code{simmr_output} object.}
}
\value{
A \code{remix_PI} object with
\item{predint}{predictive intervals for each group and tracer}
\item{resid}{residuals for each group and tracer.}
}
\description{
Plotting the consumer's isotope values against those predicted by the model
can be useful to detect model bias, or instance when consumers fall outside
the source polygon. This function also returns residuals for the consumers
as their deviations from the mean predicted value.
}
\details{
New samples should fall within the predictive intervals
with 95% (or other user-specified interval) probability.
Model bias is indicated by consumer data that fall outside the predictive
intervals. Bias may occur, among other things due to
missed sources or incorrect fractionation estimates.
}
\author{
Christopher J. Brown
}
