% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remixsiar.R
\docType{package}
\name{remixsiar}
\alias{remixsiar}
\alias{remixsiar-package}
\title{Tools for model checking of Bayesian isotope mixing
models}
\section{remixsiar vignettes}{
 
 \code{vignette('remixsiar')} for the basics
 \code{vignette('power-analysis')} for how to use this package to do power analysis
 \code{vignette('remixsiar-additional-methods')} for methods written up in appendix of the above paper.
}

