% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_sources.R
\name{rm_sources}
\alias{rm_sources}
\title{Remove similar sources to make n-1 groups based on Pianka's measure}
\usage{
rm_sources(nsources, s_means, s_sds, s_names)
}
\arguments{
\item{nsources}{\code{integer} giving number of sources}

\item{s_means}{\code{matrix} giving tracer means (columns) for each source}

\item{s_sds}{\code{matrix} giving tracer SD (columns) for each source}

\item{s_names}{\code{character} giving source names}
}
\value{
\code{list} containing grouped sources data with objects:
s_means (source means), s_sds (source SDs), grouped (two groups that were
lumped), grpids (IDs of new groups), s_names (new source names).
}
\description{
Remove similar sources to make n-1 groups based on Pianka's measure
}
\details{
identifies the two sources with highest similarity then combines
their isotopic values as the mean of the two distributions.
}
\author{
Christopher J. Brown
}
