% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Simulate consumer isotope ratios to evaluate posterior model fits}
\usage{
simdata(simmr_in, simmr_out, fast = TRUE)
}
\arguments{
\item{simmr_in}{A \code{simmr_input} object, the output of simmr_mcmc}

\item{simmr_out}{A \code{simmr_output} object}

\item{fast}{A \code{logical} indicating whether to sub-sample the
full mcmc chain, to speed simulation of new data.}
}
\value{
A \code{remix_simdata} object containing for each tracer:
\item{bias}{A \code{data.frame} with estimates of model bias - mean deviation
of the consumer data from the mean estimate}
\item{rmse}{Variance of consumer isotope ratio predictions
estimated as root-mean-square error.}
\item{xmean}{Mean predicted consumer isotope ratio}
\item{xvar}{Variance in predicted consumer isotope ratios}
}
\description{
Simulations of consumer isotope ratios are useful to evaluate
model fits for bias and precision (variance).
This function simulates data from  \code{simmr_output} objects to
estimate model bias and precision.
Multiple groups are allowed.
}
\details{
High bias is indicative of a mis-specified model, for instance
one that is missing an important source or an inaccurate fractionation.
See \code{\link{remix_PI}} and \code{\link{remix_2DPI}} for plotting
model predictions for 1 and 2 tracers.
}
\author{
Christopher J. Brown christo.j.brown@gmail.com
}
