% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_priors.R
\name{simdata_priors}
\alias{simdata_priors}
\title{Simulate consumer isotope ratios to evaluate priors}
\usage{
simdata_priors(simmr_in, simmr_priors)
}
\arguments{
\item{simmr_in}{A \code{simmr_input} object, the output of simmr_mcmc}

\item{simmr_priors}{A \code{data.frame} object containing draws
of dietary proportions from the prior distributions.}
}
\value{
A \code{remix_simdata} object containing for each tracer:
\item{bias}{A \code{data.frame} with estimates of model bias - mean deviation
of the consumer data from the mean estimate}
\item{rmse}{Variance of consumer isotope ratio predictions
estimated as root-mean-square error.}
\item{xmean}{Mean predicted consumer isotope ratio}
\item{xvar}{Variance in predicted consumer isotope ratios}
\item{xmean_bar}{Mean predicted consumer isotope ratio for the mean
dietary proportions}
\item{xvar_bar}{Variance in predicted consumer isotope ratios for the mean
predicted dietary proportions}
}
\description{
Simulations of consumer isotope ratios are useful to evaluate
model fits for bias and precision (variance).
This function evaluates simulates data from  \code{simmr_input} priors
to estimate model bias and precision.
Estimates can be compared to precision and bias from posteriors.
Multiple groups are allowed.
}
\details{
High bias is indicative of a mis-specified model, for instance
one that is missing an important source or an inaccurate fractionation.
See \code{\link{remix_PI}} and \code{\link{remix_2DPI}} for plotting
model predictions for 1 and 2 tracers.
}
\author{
Christopher J. Brown christo.j.brown@gmail.com
}
