% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr_clone.R
\name{simmr_clone}
\alias{simmr_clone}
\title{Clone a data-frame then run a simmr model}
\usage{
simmr_clone(dat, K = 20, mcmc.control, prior.control)
}
\arguments{
\item{dat}{A \code{list} containing the inputs for
\code{simmr::simmr_load} including the elements named
\code{mixes}, \code{s_names}, \code{source_means},
\code{source_sds}.}

\item{K}{A \code{numeric} integer giving the number of data-clones.}

\item{mcmc.control}{A \code{list} specifying the parameters for the
mcmc algorithm, including \code{iter}, \code{burn}, \code{thin},
\code{n.chain}. See the \code{simmr} documentation for more details.}

\item{prior.control}{A \code{list} specifying the parameters for the
models' priors, including \code{means} and \code{sds}. To
generate default priors see \code{\link{default_prior}}}
}
\value{
A \code{simmr::simmr_output} object. See the documentation
of that package for more details.
}
\description{
Data cloning can be used to obtain maximum-likelihood estimates
of parameters from Bayesian MCMC models. See details.
}
\details{
Data cloning is a legitimate method for obtaining maximum-likelihood
estimates of parameters from Bayesian models. This may be useful for
estimating prior influence (see \code{\link{remix_shrink}}) or if
frequentist statistcs are desired instead of Bayesian statistics.

It is very important to choose a large enough value of \code{K}. It is
recommended to run cloning with various K values and check parameter
estimates have converged.

This algorithm may take considerable time to run if your data-set
is large.

For more details, see:
Lele SR, Dennis B, Lutscher F. Data cloning: easy maximum likelihood
estimation for complex ecological models using Bayesian Markov chain Monte
Carlo methods. Ecology letters. 2007 Jul 1;10(7):551-63.
}
\author{
Christopher J. Brown christo.j.brown@gmail.com
}
