% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr_fitsims.R
\name{simmr_fitsims}
\alias{simmr_fitsims}
\title{Fits a simmr model to a randomly generated data-set}
\usage{
simmr_fitsims(n, tracer_means, tracer_sds, snames, smeans, ssds, sdmult = 1,
  seed = 42, mcmc.control = list(iter = 50000, burn = 2000, thin = 10,
  n.chain = 4), prior.control = NULL)
}
\arguments{
\item{n}{A \code{integer} giving the sample size to simulate.}

\item{tracer_means}{A \code{numeric} vector giving means of the mixes.}

\item{tracer_sds}{A \code{numeric} vector giving SDs of the mixes.}

\item{snames}{A \code{character} giving source names.}

\item{smeans}{A \code{numeric} vector giving means of the sources}

\item{ssds}{A \code{numeric} vector giving SDs of the mixes}

\item{sdmult}{A \code{numeric} that is multiplied by source SDs}

\item{mcmc.control}{A list giving parameters for the MCMC chains.}

\item{sdmult}{A \code{numeric} giving the random seed}

\item{priorcontrol}{A list with the means and standard deviations of the
priors.}
}
\value{
A list where the first object is a \code{simmr_input} object
giving the simulated data and the second object is a \code{simmr_ouptut}
object giving the model fit.
}
\description{
Useful for obtaining fits from a sequence of randomly simulated datasets.
For instance, you could use this function to perform a sort of broad-sense
power analysis to ask how many samples are required to obtain a specific
confidence about an animal's diet.
}
\details{
The mixes and sampled randomly from normal distributions.
Make sure the means and SDs of mixes and sources are given in the same
order (e.g. Carbon then Nitrogen for all of them).
Fixing the seed is useful if you want to compare runs across different
SDs, because the random samples will have the same pattern deviations.
See \code{simmr::simmr_load} \code{simmr::simmr_mcmc} for more details of
how the data is fit.
}
\author{
Christopher J. Brown
}
