% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squid}
\alias{squid}
\title{Squid data-set}
\format{list:
\describe{
  \item{mixes}{Carbon and nitrogen values for gulls}
  \item{s_names}{source names}
  \item{source_means}{Mean carbon and nitrogen values
 for the sources}
  \item{sources_sds}{Standard deviation of the 
 carbon and nitrogen values for the sources}
}}
\usage{
squid
}
\description{
This dataset is simulations of data based on the findings in
Miller TW, Bosley KL, Shibata J, Brodeur RD, Omori K, Emmett R. Contribution of prey 
 to Humboldt squid Dosidicus gigas in the northern California Current, 
 revealed by stable isotope analyses. 
Marine Ecology Progress Series. 2013 Mar 12:123-34.
}
\keyword{datasets}
